bgp_smart_contracts/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cytoolz==0.11.2
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
Flask==2.1.2
frozenlist==1.3.0
hexbytes==0.2.2
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.9.0
ipfshttpclient==0.8.0a2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.5.1
lru-dict==1.1.7
MarkupSafe==2.1.1
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.20.1
py-dotenv==0.1
py-solc-x==1.1.1
pycryptodome==3.14.1
pyrsistent==0.18.1
python-dotenv==0.20.0
PyYAML==6.0
requests==2.27.1
rlp==2.0.1
scapy==2.4.5
semantic-version==2.9.0
six==1.16.0
toolz==0.11.2
typing_extensions==4.2.0
urllib3==1.26.9
varint==1.0.2
web3==5.29.1
websockets==9.1
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0
